[PATCH] fix bitop warnings in parallel port generic driver
pg tried to issue bitops on int, this fixes it to use an ulong instead.

+++ drivers/block/paride/pg.c
@@ -244,7 +244,7 @@ struct pg_unit {
 	int timeout;		  /* timeout requested */
 	int status;		  /* last sense key */
 	int drive;		  /* drive */
-	int access;               /* count of active opens ... */
+	unsigned long access;     /* count of active opens ... */
 	int present;		  /* device present ? */
 	char *bufptr;
 	char name[PG_NAMELEN];	  /* pg0, pg1, ... */
